cleanUp codeBase
