+++ src/components/Modal.test.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { render, screen} from '@testing-library/react';
+import { render, screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import SpringModal from './Modal';
 
@@ -11,22 +11,20 @@ describe('SpringModal Component', () => {
   test('opens and closes modal on button click', () => {
     render(<SpringModal />);
     expect(screen.queryByText('Organization Form')).not.toBeInTheDocument();
-  
+
     userEvent.click(screen.getByTestId('buttonClick'));
-  
+
     const organizationFormText = screen.getByText(/Organization Form/i);
     expect(organizationFormText).toBeInTheDocument();
-  
+
     userEvent.click(screen.getByLabelText('close'));
-  
-    
+
     expect(organizationFormText).not.toBeInTheDocument();
   });
-  
+
   test('submits the form on button click', async () => {
     render(<SpringModal />);
-    
+
     userEvent.click(screen.getByTestId('buttonClick'));
-    
   });
 });
