+++ src/components/Modal.tsx
@@ -3,55 +3,11 @@ import Backdrop from "@mui/material/Backdrop";
 import Box from "@mui/material/Box";
 import Modal from "@mui/material/Modal";
 import Button from "@mui/material/Button";
-import { useSpring, animated } from "@react-spring/web";
-import DialogTitle from '@mui/material/DialogTitle';
-import DialogContent from '@mui/material/DialogContent';
-import IconButton from '@mui/material/IconButton';
-import CloseIcon from '@mui/icons-material/Close';
-
-interface FadeProps {
-  children: React.ReactElement;
-  in?: boolean;
-  onClick?: any;
-  onEnter?: (node: HTMLElement, isAppearing: boolean) => void;
-  onExited?: (node: HTMLElement, isAppearing: boolean) => void;
-  ownerState?: any;
-}
-
-const Fade = React.forwardRef<HTMLDivElement, FadeProps>(function Fade(
-  props,
-  ref
-) {
-  const {
-    children,
-    in: open,
-    onClick,
-    onEnter,
-    onExited,
-    ownerState,
-    ...other
-  } = props;
-  const style = useSpring({
-    from: { opacity: 0 },
-    to: { opacity: open ? 1 : 0 },
-    onStart: () => {
-      if (open && onEnter) {
-        onEnter(null as any, true);
-      }
-    },
-    onRest: () => {
-      if (!open && onExited) {
-        onExited(null as any, true);
-      }
-    },
-  });
-
-  return (
-    <animated.div ref={ref} style={style} {...other}>
-      {React.cloneElement(children, { onClick })}
-    </animated.div>
-  );
-});
+import DialogTitle from "@mui/material/DialogTitle";
+import DialogContent from "@mui/material/DialogContent";
+import IconButton from "@mui/material/IconButton";
+import CloseIcon from "@mui/icons-material/Close";
+import Fade from "./Fade"
 
 const style = {
   position: "absolute" as "absolute",
@@ -72,7 +28,12 @@ export default function SpringModal() {
 
   return (
     <div>
-      <Button onClick={handleOpen} variant="contained" sx={{ m: 50, mt: 3 }} data-testid="buttonClick">
+      <Button
+        onClick={handleOpen}
+        variant="contained"
+        sx={{ m: 50, mt: 3 }}
+        data-testid="buttonClick"
+      >
         Add Organization
       </Button>
       <Modal
@@ -94,7 +55,7 @@ export default function SpringModal() {
                 aria-label="close"
                 onClick={handleClose}
                 sx={{
-                  position: 'absolute',
+                  position: "absolute",
                   right: 8,
                   top: 8,
                   color: (theme) => theme.palette.grey[500],
@@ -103,8 +64,7 @@ export default function SpringModal() {
                 <CloseIcon />
               </IconButton>
             </DialogTitle>
-            <DialogContent dividers>
-        </DialogContent>
+            <DialogContent dividers></DialogContent>
           </Box>
         </Fade>
       </Modal>
